+++ src/com/brian/main/HUD.java
@@ -5,7 +5,7 @@ import java.awt.Graphics;
 
 public class HUD {
 
-    public static int HEALTH = 100;
+    public static float HEALTH = 100;
 
     private int score = 0;
     private int level = 1;
@@ -19,7 +19,7 @@ public class HUD {
         g.setColor(Color.gray);
         g.fillRect(15, 15, 200, 32);
         g.setColor(Color.getHSBColor((1f * HEALTH) / 360, 1f, 1f));
-        g.fillRect(15, 15, HEALTH * 2, 32);
+        g.fillRect(15, 15, (int) HEALTH * 2, 32);
         g.setColor(Color.white);
         g.drawRect(15, 15, 200, 32);
